Do not automatically connect to servers
This commit makes significant changes to how Dataspex handles remotes, both for
sending ClojureScript data off for remote inspection and for viewing JVM/remote
inspections.

The original behavior was to try to connect to localhost:7117 both for sending
inspection data and for fetching JVM renders. This caused a bunch of errors in
the console when the server wasn't up, and there was no way to connect other
host names to the extension panel - meaning there was no way to inspect multiple
JVMs.

The new behavior gives the user control over which remotes to connect. This
means JVM connections will no longer automatically appear, but the user has full
control.

This solution effectively makes it impossible to view a JVM process in the
extension without a ClojureScript build on the inspected page. The fix for this
is to provide a UI to connect remotes in the panel itself.