Remove panel-specific light/dark themes
The user can control themes with their browser. The per-panel theme selector was
buggy and didn't look good - and there was no good solution to it.

Problems with per-panel theme:

- Theme becomes branded into rendered hiccup. This means viewing data from a
  light mode environment in a dark mode environment would give you a panel in
  the wrong theme, switching it would cause the wrong theme on the other end.
- There was no way to change the theme of the document when all panels had
  changed theme.
- Occasionally the document would use one theme and panels another.

The last one could probably have been solved, but there are no good answers to
the other problems, so good riddance.