Prefer reader tags when available, and pr-str to discover them
Checking the result of pr-str is the only reliable way I know of for checking if
a value has a literal representation. Unfortunately it is not cheap, so it is
pushed back as far as possible. However, this causes JavaScript objects with
registered reader tags to be formatted as generic objects. This commit pr-strs
JavaScript objects in order to prefer their reader tag when there is one.