Change from fixed to minimum sized key columns
The consistency achieved with the fixed size key column hasn't been much of
benefit, and it wastes space for a tool that lives in a cramped environment.
Instead, add some more horizontal white-space between the columns but keep the
key column as narrow as possible, allowing it to change dimensions as you
navigate.